VisiLibity Optimization:
VisiLibity (V) is not as fast as VisibilityGraph (VG) when it comes to
finding shortest paths between two points. VG's shortest path calculation
has an optimization for the trivial case where the start and finish points
a) can see each other directly, and
b) exist in an environment with a few holes each with a few vertices.

This commit adds a similar optimization to the V implementation and
appears to close the performance gap between the two shortest path
calculation algorithms.